--- a/TrackletAlgorithm/TrackletProjectionMemory.h
+++ b/TrackletAlgorithm/TrackletProjectionMemory.h
@@ -2,7 +2,7 @@
 #define TrackletAlgorithm_TrackletProjectionMemory_h
 
 #include "Constants.h"
-#include "MemoryTemplateTPROJ.h"
+#include "MemoryTemplate.h"
 #include "globalFunctions.h"
 
 // TrackletProjectionBase is where we define the bit widths, which depend on
@@ -240,8 +240,7 @@ private:
 };
 
 // Memory definition
-template<int TProjType> using TrackletProjectionMemory = MemoryTemplateTPROJ<TrackletProjection<TProjType>, 1, kNBits_MemAddr,1>;
-//template<int TProjType> using TrackletProjectionMemory = MemoryTemplate<TrackletProjection<TProjType>, 1, kNBits_MemAddr>;
+template<int TProjType> using TrackletProjectionMemory = MemoryTemplate<TrackletProjection<TProjType>, 1, kNBits_MemAddr>;
 // FIXME: double check number of bits for bx and for memory address
 
 #endif