This PR generalizes the PC to all seeds. It also increases the default number of events to test to 100 and reverts commit 83fdaca8d86e9329e0b1a7ea1afa62c111926776, which I believe is meant for a different branch.

The results for the C-simulation of PC_L1L2C are unchanged, as is the timing. The latency and utilization of LUTs and FFs is modestly improved:
| | before | after |
| ------------- | ------------- | ------------- |
| Min. clock period | 3.141 ns | 3.141 ns |
| Max. latency | 10 cycles | 7 cycles |
| LUT | 1248 | 1239 | 
| FF | 1000 | 992 |
| DSP | 46 | 46 |
| BRAM | 4 | 4 |

There are still errors in the C-simulation of PC_L1L2C, and indeed most instances that I tested, but I don't think these are related to the generalization to all seeds.